tests: keep pytest out of the chunkers testsuite package __init__
borg.selftest imports the *_self_test.py modules from
src/borg/testsuite/chunkers/, so that package's __init__ runs on every
borg invocation - and a normal borg install has no pytest. Since
8f2de9e80 the __init__ had a hard "import pytest", needed only by the
chunker_with_kernel() helper, so a non-editable install without the dev
dependencies died at startup:

    File "borg/archiver/__init__.py", line 51, in <module>
        from ..selftest import selftest
    File "borg/selftest.py", line 24, in <module>
        from .testsuite.chunkers.buzhash_self_test import ChunkerTestCase
    File "borg/testsuite/chunkers/__init__.py", line 4, in <module>
        import pytest
    ModuleNotFoundError: No module named 'pytest'

chunker_with_kernel() moves into a separate pytest_helpers module that
only the pytest-based tests import, so the self-test modules can be
imported without pytest again.

The guard in selftest() could not catch this: "pytest" not in
dir(module) only inspects the leaf self-test module, never a package
__init__ it imports - as its own comment says.
