+++ src/borg/testsuite/chunkers/__init__.py
@@ -1,29 +1,13 @@
+# Note: this is imported by borg.selftest via the *_self_test.py modules in this package,
+#       so it must not import pytest - a normal borg install does not have it.
+#       Helpers that need pytest belong into pytest_helpers.py.
+
 import os
 import tempfile
-
-import pytest
 
 from borg.constants import *  # noqa
 
 from ...chunkers import has_seek_hole
-
-
-def chunker_with_kernel(monkeypatch, envvar, kernel, make):
-    """Build a chunker with <envvar> set to <kernel>, or skip if it cannot run here.
-
-    Which kernels exist depends on the build (compiler support) and on the CPU, so a
-    kernel that cannot be selected is not a failure - the test is skipped, carrying
-    the reason the chunker gave, so `pytest -rs` shows which kernels a machine covered.
-    Requesting an unusable kernel raising (rather than falling back) is itself tested,
-    see test_kernel_env_rejects_unusable.
-    """
-    monkeypatch.setenv(envvar, kernel)
-    try:
-        chunker = make()
-    except ValueError as err:
-        pytest.skip(str(err))
-    assert chunker.kernel == kernel  # we must have got what we asked for, not a fallback
-    return chunker
 
 
 def cf(chunks):
